// packages/frontend/src/openapi/dnd-api/model/currentPhase.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */

export type CurrentPhase = (typeof CurrentPhase)[keyof typeof CurrentPhase];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CurrentPhase = {
  preparation: "preparation",
  idle: "idle",
  action: "action",
} as const;